requirements.txt
streamlit
pandas
pyairtable
plotly
pytz
streamlit-oauth
streamlit-cookies-manager
openai
streamlit-chat
python-telegram-bot